Update main.py
Modification pour gérer header si le fichier existe déjà ou pas

+++ main.py
@@ -39,4 +39,9 @@ itemJson = {"light": 0, "temperature": 21.4, "moisture": 18, "conductivity": 373
 async def update_df(itemJson : str):
     item = json.loads(itemJson)
     df = pd.json_normalize((item))
-    df.to_csv('capteurs.csv', mode='a',header=False)
+    fileName = r"./capteurs.csv"
+    fileObj = Path(fileName)
+    if fileObj.is_file() :
+        df.to_csv('capteurs.csv',mode='a',header=False)
+    else :
+        df.to_csv('capteurs.csv',mode='a')
